Move db schema definitions to db/tables.dart
